// src/config.d.ts
export declare const config: {
    port: number;
    quotas: {
        gemini: {
            requestsPerMinute: number;
            requestsPerDay: number;
        };
        huggingface: {
            requestsPerMinute: number;
            requestsPerDay: number;
        };
        openrouter: {
            requestsPerMinute: number;
            requestsPerDay: number;
        };
        ollama: {
            requestsPerMinute: number;
            requestsPerDay: number;
        };
    };
};
//# sourceMappingURL=config.d.ts.map